Move DOM-Manipulation and Logic functions to separate modules
